fixing models, dummy data, posting script

+++ sensorgridapi/sensordata/models.py
@@ -5,9 +5,9 @@ from django.db import models
 
 # template for sensor grid data
 class SensorData(models.Model):
-    battery = models.DecimalField(max_digits=3, decimal_places=2, null=True)
+    battery = models.DecimalField(max_digits=5, decimal_places=2, null=True)
     created_at = models.DateTimeField(null=True)
-    data = models.DecimalField(max_digits=3, decimal_places=2, null=True)
+    data = models.DecimalField(max_digits=5, decimal_places=2, null=True)
     data_type = models.CharField(max_length=100, blank=True, default='')
     message_id = models.IntegerField(null=True)
     network = models.IntegerField(null=True)
@@ -18,7 +18,7 @@ class SensorData(models.Model):
     recieved_at = models.DateTimeField(auto_now_add=True)
     record_id = models.CharField(max_length=100, blank=True, default='')
     timestamp = models.DateTimeField(auto_now_add=True)
-    version = models.DecimalField(max_digits=3, decimal_places=2, null=True)
+    version = models.IntegerField(null=True)
 
     class Meta:
         ordering = ('id',)
